+++ versioneer.py
@@ -297,7 +297,7 @@ def get_root():
     """Get the project root directory.
 
     We require that all commands are run from the project root, i.e. the
-    directory that contains setup.py, setup.cfg, and versioneer.py .
+    directory that contains setup.py, setup.cfg, and versioneer.py.
     """
     root = os.path.realpath(os.path.abspath(os.getcwd()))
     setup_py = os.path.join(root, "setup.py")
@@ -362,7 +362,7 @@ def get_config_from_root(root):
 
 
 class NotThisMethod(Exception):
-    """Exception raised if a method is not valid for the current scenario."""
+    """The exception is raised if a method is not valid for the current scenario."""
 
 
 # these dictionaries contain VCS-specific tools
@@ -639,7 +639,7 @@ def git_pieces_from_vcs(tag_prefix, root, verbose, run_command=run_command):
 
     This only gets called if the git-archive 'subst' keywords were *not*
     expanded, and _version.py hasn't already been rewritten with a short
-    version string, meaning we're inside a checked out source tree.
+    version string, meaning we're inside a checked-out source tree.
     """
     GITS = ["git"]
     if sys.platform == "win32":
@@ -726,7 +726,7 @@ def git_pieces_from_vcs(tag_prefix, root, verbose, run_command=run_command):
 
 
 def plus_or_dot(pieces):
-    """Return a + if we don't already have one, else return a ."""
+    """Return a + if we don't already have one, else return a."""
     if "+" in pieces.get("closest-tag", ""):
         return "."
     return "+"